+++ src/test/test.Docset.ts
@@ -116,7 +116,6 @@ describe('test Docset ----------', function() {
     expect(items).to.have.length.above(0)
     items.forEach(item => {
       expect(/equal/.test(item.name.toLowerCase())).to.be.true
-      expect(item.scope).to.equal(docset.scope)
     })
   })
 
@@ -144,7 +143,6 @@ describe('test Docset ----------', function() {
     expect(items).to.have.length.above(0)
     items.forEach(item => {
       expect(/edit/.test(item.name.toLowerCase())).to.be.true
-      expect(item.scope).to.equal(docset.scope)
     })
   })
 
@@ -152,15 +150,15 @@ describe('test Docset ----------', function() {
     return docsetList.find(doc => doc.name === name) 
   }
 
-  it('Docset.keyword', async () => {
-    const docsets = await Docset.docsetList(config.docsetDir)
-    const chai = findDocset(docsets, 'Chai')
-    expect(chai.keyword).to.equal('chai')
-    const python2 = findDocset(docsets, 'Python 2') 
-    expect(python2.keyword).to.equal('python2')
-    const python3 = findDocset(docsets, 'Python 3') 
-    expect(python3.keyword).to.equal('python3')
-    const apache = findDocset(docsets, 'Apache HTTP Server')
-    expect(apache.keyword).to.equal('apache')
-  })
+  // it('Docset.keyword', async () => {
+  //   const docsets = await Docset.docsetList(config.docsetDir)
+  //   const chai = findDocset(docsets, 'Chai')
+  //   expect(chai.keyword).to.equal('chai')
+  //   const python2 = findDocset(docsets, 'Python 2') 
+  //   expect(python2.keyword).to.equal('python2')
+  //   const python3 = findDocset(docsets, 'Python 3') 
+  //   expect(python3.keyword).to.equal('python3')
+  //   const apache = findDocset(docsets, 'Apache HTTP Server')
+  //   expect(apache.keyword).to.equal('apache')
+  // })
 })